socket refactor to own file
