+++ server/app.js
@@ -9,10 +9,7 @@ const { json, urlencoded } = express;
 
 const app = express();
 
-const corsOptions = {
-	origin: ["http://localhost:3000"],
-	methods: ["GET", "POST", "PATCH"],
-};
+const corsOptions = require("./helpers/corsOptions");
 
 app.use(cors(corsOptions));
 app.use(logger("dev"));
@@ -21,22 +18,8 @@ app.use(urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(join(__dirname, "public")));
 
-const io = require("socket.io")({ cors: corsOptions });
+const io = require("./socket");
 app.io = io;
-
-const registerSocketAuth = require("./helpers/socket/auth");
-const registerSocketChatRooms = require("./helpers/socket/chatRooms");
-
-//functions with registered event handlers
-const onConnection = (socket) => {
-	registerSocketAuth(io, socket);
-	registerSocketChatRooms(io, socket);
-};
-
-const { socketAuth } = require("./middleware/auth");
-
-io.use(socketAuth);
-io.on("connection", onConnection);
 
 const indexRouter = require("./routes/index");
 const authRouter = require("./routes/auth");
