+++ server/helpers/corsOptions.js
@@ -1,0 +1,6 @@
+const corsOptions = {
+	origin: ["http://localhost:3000"],
+	methods: ["GET", "POST", "PATCH"],
+};
+
+module.exports = corsOptions;
